insercción de clases en service y dao

+++ src/main/kotlin/service/IProductoService.kt
@@ -2,7 +2,7 @@ package org.example.service
 
 interface IProductoService {
 
-    fun insertarProducto(nombre: String, precio: Int, stock: Int)
-    fun eliminarProducto(nombre: String, precio: Int, stock: Int)
+    fun insertarProducto(nombre: String, precio: Double, stock: Int)
+    fun eliminarProducto(nombre: String, precio: Double, stock: Int)
 
 }
